zmena velikosti  krizovatek na 15

+++ src/xroads/agents/SpawnerAgent.java
@@ -73,7 +73,7 @@ public class SpawnerAgent extends Agent {
 		addBehaviour(new SpawnWorldBehaviour(pGridWidth, pGridHeight));
 
 		// Cyklicka kontrola krizovatek
-		addBehaviour(new GuiRefreshBehaviour(this, (int) World.TIMESTEP / 3, pGridWidth, pGridHeight));
+		addBehaviour(new GuiRefreshBehaviour(this, World.TIMESTEP / 3, pGridWidth, pGridHeight));
 
 		// naslouchani na infa o stavu
 		addBehaviour(new CrossroadStatusListener(statusInformConvId));
